[lesson_26] invoices, queues, jobs, emails

+++ app/Http/Controllers/Orders/PaypalController.php
@@ -12,6 +12,8 @@ class PaypalController extends Controller
 {
     public function __invoke(Order $order)
     {
+        $this->authorize('view', $order);
+
         $order->loadMissing(['user', 'transaction', 'products']);
 
         Cart::instance('cart')->destroy();
